extracting components on extention
